\name{fixef.nlmm}
\alias{fixef.nlmm}
\alias{fixef.nlmm.counts}

\title{
Extract Generalized Mixed-Effects Models Coefficients
}
\description{
\code{fixef} extracts estimated fixed effects from \code{nlmm} objects.
}
\usage{
\method{fixef}{nlmm}(object)
}

\arguments{
  \item{object}{
an \code{nlmm} or \code{nlmm.counts} object.
}

}

\value{
a vector of estimated fixed effects. 
}

\author{
Marco Geraci
}

\seealso{
\code{\link{nlmm}}
\code{\link{summary.nlmm}}
}

\keyword{coefficients}

