\name{nlmm-internal}
\title{Internal nlmm objects}
\alias{.First.lib}
\alias{.Last.lib}

\alias{lme2nlmm}
\alias{loglik_nlmm}
\alias{loglik_alpha_nlmm}
\alias{pchibarsq}
\alias{Vchibarsq}
\alias{wchibarsq}
\alias{C_wdensgl}
\alias{C_ll}

\alias{slash}
\alias{generate.dist}
\alias{generate.design}
\alias{generate.data}

\description{Internal nlmm objects.}
\details{These are not to be called by the user.}
\keyword{internal}