\name{nlmm-package}
\alias{nlmm-package}
\docType{package}
\title{nlmm: Generalized Laplace Mixed-Effects Models}

\description{
Mixed-effects models using the generalized Laplace distribution.
}

\details{
\tabular{ll}{
Package: \tab nlmm\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2019-08-02\cr
License: \tab GPL (>=3)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marco Geraci [aut, cph, cre], Alessio Farcomeni [ctb]

Maintainer: Marco Geraci <geraci@mailbox.sc.edu>
}

\keyword{random effects}
