\name{print.nlmm}
\alias{print.nlmm}

\title{
Print an \code{nlmm} Object
}
\description{
Print an object generated by \code{\link{nlmm}}.
}
\usage{
\method{print}{nlmm}(x, digits = max(3, getOption("digits") - 3), ...)
}

\arguments{
  \item{x}{
an \code{nlmm} object.
}
  \item{digits}{
a non-null value for digits specifies the minimum number of significant digits to be printed in values.
}

  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{nlmm}}
}

\keyword{print}
