\name{print.summary.nlmm}
\alias{print.summary.nlmm}

\title{
Print an \code{nlmm} Summary Object
}
\description{
Print summary of an \code{nlmm} object.}

\usage{
\method{print}{summary.nlmm}(x, digits = max(3, getOption("digits") - 3), ...)
}

\arguments{
  \item{x}{
a \code{summary.nlmm} object.
}
  \item{digits}{
a non-null value for digits specifies the minimum number of significant digits to be printed in values.
}

  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}

\seealso{
\code{\link{nlmm}}, \code{\link{summary.nlmm}}
}


\keyword{print}
\keyword{summary}

