\name{vcov.nlmm}
\alias{vcov.nlmm}

\title{
Calculate Variance-Covariance Matrix for a Fitted Generalized Laplace Mixed-Effects Object
}

\description{
Returns the variance-covariance matrix of the all the parameters of a fitted \code{nlmm} object.
}
\usage{
\method{vcov}{nlmm}(object, ...)
}

\arguments{
  \item{object}{
an \code{nlmm} object.
}
  \item{\dots}{
not used.
}
}


\value{
a matrix.
}


\author{
Marco Geraci
}

\seealso{
\code{\link{nlmm}}
}


\keyword{covariance}
\keyword{standard errors}