\name{nlmm-internal}
\title{Internal nlmm objects}
\alias{.First.lib}
\alias{.Last.lib}

\alias{AIC.nlmm}
\alias{lme2nlmm}
\alias{loglik_nlmm}
\alias{loglik_reStruct_nlmm}
\alias{loglik_fixed_nlmm}
\alias{loglik_alpha_nlmm}
\alias{pchibarsq}
\alias{Vchibarsq}
\alias{wchibarsq}
\alias{C_wdensgl}
\alias{C_ll}

\description{Internal nlmm objects.}
\details{These are not to be called by the user.}
\keyword{internal}