\name{nlmm-package}
\alias{nlmm-package}
\docType{package}
\title{nlmm: Generalized Laplace Mixed-Effects Models}

\description{
The nlmm package provides functions to fit linear mixed models
	based on convolutions of the generalized Laplace (GL) distribution.
	The GL mixed-effects model includes four special cases with normal random
	effects and normal errors (NN), normal random effects and Laplace errors (NL),
	Laplace random effects and normal errors (LN), and Laplace random effects
	and Laplace errors (LL). The methods are described in Geraci and Farcomeni (2020).
	See also Geraci (2017) for details on special cases.
}

\details{
\tabular{ll}{
Package: \tab nlmm\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2020-03-21\cr
License: \tab GPL (>=3)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marco Geraci [aut, cph, cre], Alessio Farcomeni [ctb]

Maintainer: Marco Geraci <geraci@mailbox.sc.edu>
}

\references{
Geraci M (2017). Mixed-effects models using the normal and the Laplace distributions: A 2 x 2 convolution scheme for applied research. arXiv:1712.07216v1 [stat.ME]. URL: \url{https://arxiv.org/abs/1712.07216v1}.

Geraci, M. and Farcomeni A (2020). A family of linear mixed-effects models using the generalized Laplace distribution. Statistical Methods in Medical Research, \url{https://doi.org/10.1177/0962280220903763}.
}

\keyword{random effects}
